+++ NEWS.md
@@ -1,3 +1,19 @@
+# 0.5.0
+
+* improve multicore support, pbapply is now used to split up the calculation of points of stability; this is much faster than with futures but does not work on Windows
+* give the option to set the corridor manually via lower_limit and upper_limit, which will override precision parameters
+* change parameter precision to precision_absolute and precision_rel to precision_relative, if you used these parameters before, you need to change your code
+* parameter rhos is no rho
+* precision_relative now takes a numeric value, instead of being a logical
+* change how corridor limits are set up, now closures are used to handle all cases
+* several improvements in documentation, adding more details and hints
+* several improvements in the readme, especially for the benchmarks
+* linting the package with lintr
+* deprecate old precision parameters (preparing for version 0.6.0) and rhos
+* improve parameter order
+* use scientific notation more consistently
+* use roxygen md
+
 # fastpos 0.4.1
 
 * tests now show more information if they fail, also the tolerance for an